style: create global buttons file
