+++ plugin-hrm-form/src/components/transfer/TransferButton.tsx
@@ -19,7 +19,7 @@ import React from 'react';
 import { Actions, Template, TaskContextProps, withTaskContext } from '@twilio/flex-ui';
 
 import { canTransferConference } from '../../utils/transfer';
-import { TransferStyledButton } from '../../styles/HrmStyles';
+import { TransferStyledButton } from '../../styles/buttons';
 import HhrTheme from '../../styles/HrmTheme';
 
 const TransferButton: React.FC<TaskContextProps> = ({ task }) => {
